+++ lib/pages/home/home.dart
@@ -4,6 +4,7 @@ import 'package:trackless/global/app_state.dart';
 import 'package:trackless/components/async_progress.dart';
 import 'package:trackless/components/list_work/list_work.dart';
 import 'package:trackless/pages/home/home_refresh_action.dart';
+import 'package:trackless/pages/work_dialog/work_dialog.dart';
 import 'package:trackless/trackless/trackless_failure.dart';
 import 'package:trackless/trackless/trackless_location.dart';
 import 'package:trackless/trackless/trackless_work.dart';